2.1.1:
Fixed an oversight in the documentation which did not explain that the use of `addFields` is preferred over specifying `fields` in its entirety.

Also made sure the set of fields that are directly queried for exact matches to speed up queries are always indexed, even if they are left out of the `fields` option.

If you received no search results due to a combination of these issues, run the reindex task again and try those searches again.